requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
httpx
mcp
anthropic
selenium
cloudinary
